@@ -15,22 +15,22 @@ long_description = """
 See https://github.com/DLR-RM/rl-baselines3-zoo
 """
 install_requires = [
-    "sb3_contrib>=2.4.0a4,<3.0",
-    "gymnasium~=0.29.1",
+    "sb3_contrib>=2.6.0,<3.0",
+    "gymnasium>=0.29.1,<1.2.0",
     "huggingface_sb3>=3.0,<4.0",
     "tqdm",
     "rich",
     "optuna>=3.0",
     "pyyaml>=5.1",
     "pytablewriter~=1.2",
+    "shimmy~=2.0",
 ]
-# TODO(antonin): update to rliable>=1.1.0 once PR is merged and released
-plots_requires = ["seaborn", "rliable @ git+https://github.com/araffin/rliable@patch-1", "scipy~=1.10"]
+plots_requires = ["seaborn", "rliable~=1.2.0", "scipy~=1.10"]
 test_requires = [
     # for MuJoCo envs v4:
-    "mujoco~=2.3",
+    "mujoco>=2.3,<4",
     # install parking-env to test HER
-    "highway-env==1.8.2",
+    "highway-env>=1.10.1,<1.11.0",
 ]
 
 setup(
@@ -56,7 +56,7 @@ setup(
     long_description=long_description,
     long_description_content_type="text/markdown",
     version=__version__,
-    python_requires=">=3.8",
+    python_requires=">=3.9",
     # PyPI package information.
     project_urls={
         "Code": "https://github.com/DLR-RM/rl-baselines3-zoo",
@@ -68,10 +68,10 @@ setup(
     },
     classifiers=[
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
     ],
 )
 
