@@ -3,7 +3,6 @@ Run multiple experiments on a single machine.
 """
 
 import subprocess
-from typing import List
 
 import numpy as np
 
@@ -33,6 +32,6 @@ for algo in ALGOS:
                     "-f",
                     log_folder,
                 ]
-                arg_str_list: List[str] = list(map(str, args))
+                arg_str_list: list[str] = list(map(str, args))
 
                 ok = subprocess.call(["python", "train.py", *arg_str_list])
