@@ -32,7 +32,12 @@ def train() -> None:
     )
     parser.add_argument("-n", "--n-timesteps", help="Overwrite the number of timesteps", default=-1, type=int)
     parser.add_argument("--num-threads", help="Number of threads for PyTorch (-1 to use default)", default=-1, type=int)
-    parser.add_argument("--log-interval", help="Override log interval (default: -1, no change)", default=-1, type=int)
+    parser.add_argument(
+        "--log-interval",
+        help="Override log interval (default: -1, no change, -2: no logging useful when using custom logging freq)",
+        default=-1,
+        type=int,
+    )
     parser.add_argument(
         "--eval-freq",
         help="Evaluate the agent every n steps (if negative, no evaluation). "
@@ -83,7 +88,7 @@ def train() -> None:
         help="Sampler to use when optimizing hyperparameters",
         type=str,
         default="tpe",
-        choices=["random", "tpe", "skopt"],
+        choices=["random", "tpe", "auto"],
     )
     parser.add_argument(
         "--pruner",
@@ -104,6 +109,7 @@ def train() -> None:
         "--storage", help="Database storage path if distributed optimization should be used", type=str, default=None
     )
     parser.add_argument("--study-name", help="Study name for distributed optimization", type=str, default=None)
+    parser.add_argument("--trial-id", help="Trial id to load for a Optuna storage.", type=int)
     parser.add_argument("--verbose", help="Verbose mode (0: no output, 1: INFO)", default=1, type=int)
     parser.add_argument(
         "--gym-packages",
@@ -260,6 +266,7 @@ def train() -> None:
         device=args.device,
         config=args.conf_file,
         show_progress=args.progress,
+        trial_id=args.trial_id,
     )
 
     # Prepare experiment and launch hyperparameter optimization if needed
